Authentication now works with client version 3.3.3a, THANKS to Unix.
Unix reminded me that Blizz changed the key for Encrypt/Decrypt to use the 40 byte K instead of the 20 byte Hash it used to use.  Spurious was using K, but when it wrote out to the Accounts Table, it was only writing out 20 bytes and reading back in 20 bytes when it needed it.  Now it correctly writes and reads 40 bytes.

It is hitting an error on the Character Enum packet now, I have not tracked down exactly what it is, I suspect something to do with the items maybe.