Adjusted number reputations for character from 64 to 128 per ArcEmu.
This will most likely break existing characters due to there not being enough reputation flag/value pairs in character_reputation in the database.

I am unable to test this at the moment, but in theory that is what would happen.

We still need to track down the arithmetic overflow error in the setfaction routine.