Did some work on Addon Info packet to fix Blizzard Addons being disabled.
Keybindings are not set, so for now you have to go to the keybindings when you login and reset to default before you can move or use any keys.

I disabled the creature spawns, there is a problem with them somewhere in the update packet I think.  You can enter world this way.

Gameobjects spawn and stayed spawned as long as you do not move, once you move they are despawned.

This is turning into a full time job. :D