Initial Spell Packets, Gameobject spawns and so on.
Fixed the gossip calls in the Spirit Healer script.

Thanks to UniX the initial spell packet has been corrected.  Chat does not give the error that you do not know the language, but it still does not work.

Unremarked the Game Objects spawns, UniX helped with this also.  Gameobjects now spawn, and seem to stay spawned when the player moves.  I changed the Pitch Value on the movement constants to match what Mangos is using.

Played around some with Creature Spawns, but they crash the client with #132 error at the moment, too tired to look into it further this evening.

UniX also fixed the Action Buttons, so they are now showing on the action bar.  Thanks UniX.

May have changed some other things in playing around, but don't remember.